chore: move files to be organized in a DDD way
- Tried to place the files organized by Domain Driven Design
