+++ src/modules/users/services/AuthenticateUserService.ts
@@ -2,11 +2,11 @@ import { getRepository } from 'typeorm';
 import { compare } from 'bcryptjs';
 import { sign } from 'jsonwebtoken';
 
-import AppError from '../errors/AppError';
+import authConfig from '@config/auth';
 
-import authConfig from '../config/auth';
+import AppError from '@shared/errors/AppError';
 
-import User from '../models/User';
+import User from '../infra/typeorm/entities/User';
 
 interface RequestDTO {
   email: string;
